feat: rename to match react standard
